Added support for Windows/Linux Unibody style
These changes fix the color/contrast problems with the Unibody interface in Windows/Linux, making all themes usable on those platforms.

Additionally the cursor is now visible in `source mode` in Windows/Linux.

These changes are a quick fix and the UI colors in Unibody style still need to be tweaked.